requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.14
django-countries==7.3.2
django-extensions==3.2.0
djangorestframework==3.13.1
ipython==8.4.0
pytz==2022.1
shortuuid==1.0.9
sqlparse==0.4.2
